# Conflicts:
#	Assets/BitGem.meta
#	Assets/Imported Assets/Beautify/Builtin/BeautifyBuiltinPipeline.unitypackage.meta
#	Assets/Imported Assets/Beautify/PPSv2/BeautifyPPSv2.unitypackage.meta
#	Assets/Imported Assets/Beautify/URP/Beautify_URP.unitypackage.meta
#	Assets/Imported Assets/Bitgem/Dungeon_Set_01/dungeon_set_01.meta
#	Assets/Imported Assets/Bitgem/Scenes/skeleton_crew_scene.unity.meta
#	Assets/Imported Assets/Bitgem/Skeleton_Archer/Materials.meta
#	Assets/Imported Assets/FantasySpellsEffectsPack/Patch/DistortPatchForHDRP&URP.unitypackage.meta
#	Assets/Imported Assets/FantasySpellsEffectsPack/Patch/GammaColorSpace.unitypackage.meta
#	Assets/Imported Assets/FantasySpellsEffectsPack/Patch/LinearColorSpace.unitypackage.meta
#	Assets/Imported Assets/Kevin Iglesias/Melee Warrior Animations/Animations/TwoHanded/ShokeAttack2.anim.meta
#	Assets/Plugins/Demigiant/DOTween.meta
#	Assets/Resources/Animation/Animator/Monster/SkeletonPase3.controller.meta
#	Assets/Resources/DunGen Settings.asset
#	Assets/Resources/Prefab/Effect/SkillEffect/Monster/Attacks/SkeletionShokeWaveEffect.prefab.meta
#	Assets/Scripts/Etc/Item/Weapon/Sword.cs

+++ Assets/Scripts/UI/WeaponMasteryUI/ExpBar.cs
@@ -11,7 +11,7 @@ public class ExpBar : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        float ratio = (WeaponManager.Instance._weaponDic[name].exp / WeaponManager.Instance._weaponDic[name].expMax);
+        float ratio = ((float) WeaponManager.Instance._weaponDic[name].exp / (float)WeaponManager.Instance._weaponDic[name].expMax);
         if (image)
             image.fillAmount = ratio;
     }
